Contacts_partial string find
We're going to make our own Contacts application! The application must perform two types of operations:
add name, where name is a string denoting a contact name. This must store name as a new contact in the application.
find partial, where partial is a string denoting a partial name to search the application for. It must count the number of contacts starting with partial and print the count on a new line.

Given n sequential add and find operations, perform each operation in order.

Sample Input
4
add hack
add hackerrank
find hac
find hak

Sample Output
2
0

Explanation
We perform the following sequence of operations:
Add a contact named hack.
Add a contact named hackerrank.
Find and print the number of contact names beginning with hac. There are currently two contact names in the application and both of them start with hac, so we print 2 on a new line.
Find and print the number of contact names beginning with hak. There are currently two contact names in the application but neither of them start with hak, so we print 0 on a new line.

Actually this commit has solution, maybe a simple one however, for the autocomplete use case. It's interesting to realize that dictionary and string slicing could be used to initialize partial string finding. What does that mean? Consider the autocomplete scenario we face while typing on Gmail or LinkedIn message box. How is that done? Well as simply as extracting a combination of letters from the given word and forming a dictionary with those letters. So when we type a few letters in a word an algorithm automatically pulls the key in the dictionary which is the word. Of course to scale with speed we need powerful techniques like machine learning and deep learning.